+++ src/templates/styledFileReactNative.ts
@@ -1,0 +1,8 @@
+export default () => (
+`import styled from 'styled-components/native';
+
+export const Container = styled.SafeAreaView\`
+
+\`;
+`
+);