+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-All notable changes to the Create React TSX Component extension will be documented in this file.
+All notable changes to the Create React and React Native TSX Component extension will be documented in this file.
+
+## [1.2.0]
+
+- Added support for creating components for React Native.
 
 ## [1.0.1]